boop

+++ src/felix86/common/signal_queue.hpp
@@ -21,7 +21,7 @@ struct SignalQueue {
     }
 
     PendingSignal pop() {
-        for (int i = 0; i < signal_queues.size(); i++) {
+        for (size_t i = 0; i < signal_queues.size(); i++) {
             if (!signal_queues[i].empty()) {
                 total_signals--;
                 return signal_queues[i].pop();
@@ -49,7 +49,7 @@ private:
         }
 
         PendingSignal pop() {
-            ASSERT_MSG(index - 1 >= 0, "Signal queue underflow somehow?");
+            ASSERT_MSG((int)index - 1 >= 0, "Signal queue underflow somehow?");
             PendingSignal signal = data[index - 1];
             index--;
             return signal;
@@ -60,10 +60,10 @@ private:
         }
 
     private:
-        int index = 0;
+        size_t index = 0;
         std::array<PendingSignal, 10> data{};
     };
 
-    int total_signals = 0;
+    size_t total_signals = 0;
     std::array<SignalQueueSingle, 32> signal_queues{};
 };
